TIKA-4864 - hoist the plugins-dir resolution into tika-pipes-core, shared by tika-server, PipesForkParser and the async CLI
